requirements.txt
python-telegram-bot==20.8
google-generativeai==0.4.0
python-binance
python-dotenv==1.0.1
